# openapi.yaml
openapi: 3.0.0
tags:
  - name: AIBrush

info:
  title: AIBrush API
  description: Orchestration backend for AI Brush
  version: 1.0.0
  contact:
    email: wolfgangmeyers@gmail.com
servers:
  - url: https://qb1eyw689j.execute-api.us-east-1.amazonaws.com/Prod
  - url: /
paths:
  /jobs:
    get:
      description: Returns a list of jobs.
      operationId: listJobs
      tags:
        - AIBrush
      responses:
        '200':    # status code
          description: A list of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
    post:
      description: Create a new job
      operationId: createJob
      tags:
        - AIBrush
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      description: Get a single job
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      description: Delete a job
      operationId: deleteJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "204":
          description: Success
  /jobs/{id}/cancel:
    post:
      description: Cancel a job
      operationId: cancelJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "204":
          description: Job cancelled
  /jobs/{id}/results:
    get:
      description: Get a list of results for a job
      operationId: getJobResults
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResultList"
    post:
      description: Submit a job result
      operationId: submitJobResult
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobResultInput"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResult"
  /job-results/{id}:
    get:
      description: Get a job result by id
      operationId: getJobResult
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResult"
    put:
      description: Saves a job result as a new image
      operationId: saveJobResult
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "204":
          description: Success
    delete:
      description: Deletes a job result
      operationId: deleteJobResult
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - AIBrush
      responses:
        "204":
          description: Success

components:
  schemas:
    Job:
      properties:
        id:
          type: string
        created:
          type: integer
        phrases:
          type: array
          items:
            type: string
        parent:
          type: string
        label:
          type: string
        count:
          type: integer
        iterations:
          type: integer
        cancelled:
          type: boolean
    JobInput:
      required:
        - phrases
        - count
        - iterations
        - label
      properties:
        phrases:
          type: array
          items:
            type: string
        parent:
          type: string
        label:
          type: string
        count:
          type: integer
        iterations:
          type: integer
    JobList:
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
    JobResult:
      properties:
        id:
          type: string
        job_id:
          type: string
        encoded_image:
          type: string
        encoded_latents:
          type: string
        phrases:
          type: array
          items:
            type: string
        created:
          type: integer
    JobResultList:
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/JobResult"
    JobResultInput:
      properties:
        encoded_image:
          type: string
        encoded_latents:
          type: string
      required:
        - encoded_image
        - encoded_latents
    Image:
      properties:
        id:
          type: string
        encoded_image:
          type: string
        encoded_latents:
          type: string
        phrases:
          type: array
          items:
            type: string
        created:
          type: integer